requirements.txt
# Core Dependencies
selenium==4.15.2
pandas>=1.5.0
requests>=2.28.0
beautifulsoup4>=4.11.0
numpy>=1.21.0

# AI/ML Models
transformers>=4.21.0
torch>=1.12.0

# Environment Management
python-dotenv>=0.19.0

# Optional: For enhanced functionality
# Uncomment if needed
# scikit-learn>=1.0.0
# scipy>=1.7.0
